feat: add mutations abstraction to deal with mnesia updates
The goal of this change is to be able to isolate mnesia updates into
mutations which will all be ran in a single transaction and in the
least amount of time possible.

By making first-class mutations we can also debug easily what changes
any given block or transaction makes.

This also fixes random errors that come up inside mnesia transactions:
mnesia transactions do not work properly when exceptions are thrown
inside them, even if they are captured.

closes #331