refactor: use declarative state for executing mutations
The goal of these changes is to be able to use a interchangeable state
that can perform these executions. The functional way of dealing with
any sort of state mutations allows this, and which will also allow
performing in-memory updates to the state.

This is still a WIP and still requires two things:

* Transform the current State module into Db and create a new State
  protocol that has the basic state operations
* Add tests