fix: change the name of the genesis hardfork to roma
The first hardfork is called roma in the node, and genesis in the mdw.
This is causing the mdw to return 0 for the total balances of the roma
accounts, since it only knows about genesis accounts, and does not
recognize a hardfork called roma.

The above issue is causing the reported total token supply to be less
than the actual total token supply by 89451376822397976634367408, which
is the sum of balances of all genesis accounts.

In order to fix the above problem, this commit renames genesis hardfork
to roma in the mdw, and bans the access to accounts of unrecognized
protocols in order to mitigate this kind of issues in the future.

It also adds a migration for updating mdw entries to match the correct
total token supply.